refactor(header left): added component
- added reusable header left component to reduce duplication
